refactor(src/eth-swr): move ETH SWR higher in the dapp loading to be used in dxvote
